backend/requirements.txt
fastapi>=0.111.0
python-socketio>=5.11.2
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
fish-audio-sdk>=0.1.13
ffmpeg-python>=0.2.0
httpx>=0.27.0
